+++ backend/jobplatform/src/main/java/jobplatform/fo/user/domain/entity/MemberEntity.java
@@ -29,7 +29,7 @@ public class MemberEntity {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "mbr_sq", nullable = false)
-    private int mbrSq;
+    private Long mbrSq;
 
     // 소셜_유형_코드
     @Column(name = "scl_ctgry_cd", nullable = true, columnDefinition = "varchar(20)")
@@ -79,7 +79,7 @@ public class MemberEntity {
 
     // 등록_회원_순번
     @Column(name = "insrt_mbr_sq", nullable = false)
-    private int insrtMbrSq;
+    private Long insrtMbrSq;
 
     // 등록_일시
     @CreationTimestamp // 해당 필드가 DB에 삽입될 때 현재 시간으로 자동 설정됨
@@ -88,7 +88,7 @@ public class MemberEntity {
 
     // 수정_회원_순번 --널값허용하지 않아 Integer로 수정
     @Column(name = "updt_mbr_sq", nullable = true)
-    private Integer updtMbrSq;
+    private Long updtMbrSq;
 
     // 수정_일시
     @LastModifiedDate // 해당 필드가 DB에 저장되거나 업데이트될 때마다 현재 시간으로 업데이트
